A working demo of the tree
This demo is modified from a piece of sample code(https://bl.ocks.org/d3noob/8375092), using html, css, and d3.js version 3. It's fully functional, but I do not recommend reading the source code directly, because I didn't remove a lot of unnessasary codes which are not fully understanded.